Changed variants generation to single file
