+++ libraries/tokens/components/variants.ts
@@ -1,0 +1,126 @@
+export const buttons = {
+  dark_theme: ['false', 'true'],
+  color: ['primary', 'secondary', 'danger'],
+  variant: ['outlined', 'ghost', 'ghost_icon', 'contained'],
+  state: ['enabled', 'loading', 'focus', 'hover', 'pressed', 'disabled'],
+  icon: ['none', 'leading', 'trailing'],
+}
+
+export const buttons_toggle = {
+  dark_theme: ['false', 'true'],
+  alignment: ['left', 'right', 'center'],
+  state: ['pressed', 'hover', 'focus', 'enabled'],
+  active: ['true', 'false'],
+  icon: ['false', 'true'],
+}
+
+export const chips = {
+  dark_theme: ['false', 'true'],
+  avatar: ['false', 'true'],
+  custom_icon: ['true', 'false'],
+  remove_icon: ['true', 'false'],
+  state: ['enabled', 'disabled', 'focus', 'hover', 'pressed'],
+  selected: ['false', 'true'],
+  error: ['false', 'true'],
+}
+
+export const divider = {
+  dark_theme: ['false', 'true'],
+  size: ['small', 'medium'],
+}
+
+export const navigation__breadcrumbs = {
+  dark_theme: ['false', 'true'],
+  state: ['enabled', 'focus', 'hover'],
+}
+
+export const navigation__pagination = {
+  dark_theme: ['false', 'true'],
+  range_and_total: ['true', 'false'],
+  range_select: ['true', 'false'],
+}
+
+export const navigation__tabs = {
+  dark_theme: ['false', 'true'],
+  state: ['enabled', 'focus', 'hover', 'pressed', 'disabled'],
+  active: ['true', 'false'],
+}
+
+export const navigation__top_bar = {
+  dark_theme: ['false', 'true'],
+  with_navigation: ['false', 'true'],
+}
+
+export const notifications__banner = {
+  dark_theme: ['false', 'true'],
+  variant: ['text', 'actions', 'text__icon'],
+}
+
+export const notifications__dialog = {
+  dark_theme: ['false', 'true'],
+  variant: [
+    'placeholder__actions',
+    'placeholder',
+    'scrollable_placeholder__actions',
+    'text__actions',
+    'no_title__text___actions',
+  ],
+}
+
+export const notifications__snackbar = {
+  dark_theme: ['false', 'true'],
+}
+
+export const popover = {
+  dark_theme: ['false', 'true'],
+  caret_direction: [
+    '↙︎_left_bottom',
+    '←_left',
+    '↖︎_left_top',
+    '↘︎_right_bottom',
+    '→_right',
+    '↗︎_right_top',
+    '↗︎_top_right',
+    '↑_top',
+    '↖︎_top_left',
+    '↘︎_bottom_right',
+    '↓_bottom',
+    '↙︎_bottom_left',
+  ],
+}
+
+export const stepper = {
+  dark_theme: ['false', 'true'],
+  type: [
+    'horizontal_numeric',
+    'vertical_numeric',
+    'horizontal_icon',
+    'vertical_icon',
+  ],
+  state: ['finished', 'in_progress', 'waiting'],
+}
+
+export const text_fields = {
+  dark_theme: ['false', 'true'],
+  state: [
+    'enabled',
+    'active',
+    'filled',
+    'read_only',
+    'disabled',
+    'focus',
+    'hover',
+  ],
+  multiline: ['false', 'true'],
+  validation: ['none', 'error', 'warning', 'success'],
+}
+
+export const time_picker = {
+  dark_theme: ['false', 'true'],
+  state: ['enabled', 'disabled', 'hover', 'pressed', 'selected'],
+}
+
+export const toolbar = {
+  dark_theme: ['false', 'true'],
+  fixed: ['true', 'false'],
+}
