--- libraries/tokens/components/variants/buttons.ts
@@ -1,7 +1,0 @@
-export const buttons = {
-  dark_theme: ['false', 'true'],
-  color: ['primary', 'secondary', 'danger'],
-  variant: ['outlined', 'ghost', 'ghost_icon', 'contained'],
-  state: ['enabled', 'loading', 'focus', 'hover', 'pressed', 'disabled'],
-  icon: ['none', 'leading', 'trailing'],
-}
