--- libraries/tokens/components/variants/buttons_toggle.ts
@@ -1,7 +1,0 @@
-export const buttons_toggle = {
-  dark_theme: ['false', 'true'],
-  alignment: ['left', 'right', 'center'],
-  state: ['pressed', 'hover', 'focus', 'enabled'],
-  active: ['true', 'false'],
-  icon: ['false', 'true'],
-}
