--- libraries/tokens/components/variants/chips.ts
@@ -1,9 +1,0 @@
-export const chips = {
-  dark_theme: ['false', 'true'],
-  avatar: ['false', 'true'],
-  custom_icon: ['true', 'false'],
-  remove_icon: ['true', 'false'],
-  state: ['enabled', 'disabled', 'focus', 'hover', 'pressed'],
-  selected: ['false', 'true'],
-  error: ['false', 'true'],
-}
