--- libraries/tokens/components/variants/divider.ts
@@ -1,4 +1,0 @@
-export const divider = {
-  dark_theme: ['false', 'true'],
-  size: ['small', 'medium'],
-}
