--- libraries/tokens/components/variants/navigation__breadcrumbs.ts
@@ -1,4 +1,0 @@
-export const navigation__breadcrumbs = {
-  dark_theme: ['false', 'true'],
-  state: ['enabled', 'focus', 'hover'],
-}
