--- libraries/tokens/components/variants/navigation__pagination.ts
@@ -1,5 +1,0 @@
-export const navigation__pagination = {
-  dark_theme: ['false', 'true'],
-  range_and_total: ['true', 'false'],
-  range_select: ['true', 'false'],
-}
