--- libraries/tokens/components/variants/navigation__tabs.ts
@@ -1,5 +1,0 @@
-export const navigation__tabs = {
-  dark_theme: ['false', 'true'],
-  state: ['enabled', 'focus', 'hover', 'pressed', 'disabled'],
-  active: ['true', 'false'],
-}
