--- libraries/tokens/components/variants/navigation__top_bar.ts
@@ -1,4 +1,0 @@
-export const navigation__top_bar = {
-  dark_theme: ['false', 'true'],
-  with_navigation: ['false', 'true'],
-}
