--- libraries/tokens/components/variants/notifications__banner.ts
@@ -1,4 +1,0 @@
-export const notifications__banner = {
-  dark_theme: ['false', 'true'],
-  variant: ['text', 'actions', 'text__icon'],
-}
