--- libraries/tokens/components/variants/notifications__dialog.ts
@@ -1,10 +1,0 @@
-export const notifications__dialog = {
-  dark_theme: ['false', 'true'],
-  variant: [
-    'placeholder__actions',
-    'placeholder',
-    'scrollable_placeholder__actions',
-    'text__actions',
-    'no_title__text___actions',
-  ],
-}
