--- libraries/tokens/components/variants/notifications__snackbar.ts
@@ -1,3 +1,0 @@
-export const notifications__snackbar = {
-  dark_theme: ['false', 'true'],
-}
