--- libraries/tokens/components/variants/popover.ts
@@ -1,17 +1,0 @@
-export const popover = {
-  dark_theme: ['false', 'true'],
-  caret_direction: [
-    '↙︎_left_bottom',
-    '←_left',
-    '↖︎_left_top',
-    '↘︎_right_bottom',
-    '→_right',
-    '↗︎_right_top',
-    '↗︎_top_right',
-    '↑_top',
-    '↖︎_top_left',
-    '↘︎_bottom_right',
-    '↓_bottom',
-    '↙︎_bottom_left',
-  ],
-}
