--- libraries/tokens/components/variants/stepper.ts
@@ -1,10 +1,0 @@
-export const stepper = {
-  dark_theme: ['false', 'true'],
-  type: [
-    'horizontal_numeric',
-    'vertical_numeric',
-    'horizontal_icon',
-    'vertical_icon',
-  ],
-  state: ['finished', 'in_progress', 'waiting'],
-}
