--- libraries/tokens/components/variants/text_fields.ts
@@ -1,14 +1,0 @@
-export const text_fields = {
-  dark_theme: ['false', 'true'],
-  state: [
-    'enabled',
-    'active',
-    'filled',
-    'read_only',
-    'disabled',
-    'focus',
-    'hover',
-  ],
-  multiline: ['false', 'true'],
-  validation: ['none', 'error', 'warning', 'success'],
-}
