--- libraries/tokens/components/variants/time_picker.ts
@@ -1,4 +1,0 @@
-export const time_picker = {
-  dark_theme: ['false', 'true'],
-  state: ['enabled', 'disabled', 'hover', 'pressed', 'selected'],
-}
