--- libraries/tokens/components/variants/toolbar.ts
@@ -1,4 +1,0 @@
-export const toolbar = {
-  dark_theme: ['false', 'true'],
-  fixed: ['true', 'false'],
-}
